requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
numpy==1.24.3
tensorflow==2.13.0
pymongo==4.6.0
Pillow==10.0.0
pydantic==2.5.0
pydantic-extra-types==2.4.1
bcrypt==4.1.1
python-multipart==0.0.6
